Clean up SharedAccessSignature files for the hub service so that each class has a distinct responsibility
