--- iothub/service/src/Authentication/Security/ISharedAccessSignatureCredential.cs
@@ -1,23 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.Azure.Devices
-{
-    /// <summary>
-    /// Credential interface used for authentication and authorization.
-    /// </summary>
-    internal interface ISharedAccessSignatureCredential
-    {
-        /// <summary>
-        /// Indicates if the token has expired.
-        /// </summary>
-        bool IsExpired();
-
-        /// <summary>
-        /// The date and time of expiration.
-        /// </summary>
-        DateTime ExpiryTime();
-    }
-}
